Add design reference filter
This adds a new reference filter to recognise DesignManagement::Design
instances. This is added to the GFM pipeline, so is universally
available, but it is only enabled when the appropriate feature flag is
enabled. See ee/lib/banzai/filter/design_reference_filter.rb

To enable this filter, DesignManagement itself must be enabled, and then
the following feature flag must be set:

  design_management_reference_filter_gfm_pipeline