Look for LFS objects from the fork and source
While backfilling LfsObjectsProject is ongoing, we should still
be looking at both the fork and source when finding for LFS
objects.

This way, older forks that are dependent on finding LFS objects
from the source, will still be able to find them. Newer forks,
should be able to find the object directly.