Generate smaller image sizes for Designs
Adds a new DesignUploader to generate and serve smaller versions of
design files.

The original (full-sized) design files are stored in Git LFS, and so
have a different uploader, `LfsObjectUploader`.

The decision to make a separate uploader to handle just the smaller
image versions came out of a discussion in
https://gitlab.com/gitlab-org/gitlab/merge_requests/20511. Initially,
the direction taken was to make LfsObjectUploader conditionally generate
smaller images for LfsObject files. The feedback was that this was
adding logic that was too specific to design management into an uploader
that is very general, and that we should use a separate uploader to
handle just the smaller image versions.

https://gitlab.com/gitlab-org/gitlab/issues/12577
https://gitlab.com/gitlab-org/gitlab/issues/13815