Explore Projects default to All tab
For instances without public projects or much activity, there can be no
trending projects. Because the Explore item in the navbar led to the
Trending tab by default, some users were being confused by the lack of
data. This changes the Explore feature to default to the All tab
instead.