Blob content component
Set up the shared blob viewers lib

For the sake of re-usability, the blob content component is designed
to be a a light-weight container that connects different viewers,
based on the viewer's props. The viewers are put into the shared
folder so that more applications could benefit from using those

Smooth navigation to the linked line

Whenever a hash pointing to a line is found in the URL ,we will
smoothly scroll to the corresponding line. The same when user
explicitely clicks any line: we smoothly scroll to the line