Add Selective Sync Namespaces to Vue Node Form
This changes is broken down from a very large MR:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/22719

The overall goal is to convert the Geo Node Form from HAML to Vue.
Based on MVC, splitting this into smaller MRs is more feasible.

This MR adds the final fields to the Geo From in Vue:
  - Selective Sync Namespaces (Multi-Select Dropdown)

After this MR is merged, the submit actions and API will be added.