Generic Confirm Modal in JS
Currently, we cannot call GlModal from HAML very easily.
Throughout the backend there are countless uses of the browser
default confirm box.

This new component will be used to open the door to replace all
those instances of browser confirms with this.

An example of this being done can be found here:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/24808