Add initial tests for first-person pronouns
This commit adds initial first-person tests for first-person
pronouns, which the GitLab style guide prohibits. The test for
'I' is separate from the other tests, because a case-insensitive
test using standard word boundaries will match both 'I/O' (which is
correct usage) and 'i.e.' (which is wrong but for different
reasons).