Generalize Geo Replicable UI Components
Currently, the UI components for Geo Replicables
is isolated to Designs Only.
As we move forward, it will be easier if we have a
generalized place in the UI to house different
replicable data types.

This MR takes the Geo Designs components and
renames everything to Geo Replicable Items.

This also adds a parameterized route in the API
to support different replicable data type queries.