Ensure RepositoryLinkFilter handles Gitaly failures gracefully
We do this by setting all paths to `:raw` type if we get either
`GRPC::Unavailable` or `GRPC::DeadlineExceeded` when calling
`GitalyClient`. This is acceptable because all files existing in a
repository can be viewed in "raw" mode. For example text would be shown
on the browser, videos would be downloaded, etc.