Internationalize the renewal banner
I'll be working in this area and decided to backfill the test suite so
it'd be more easily refactorable for the feature I'm working on. In the
process I also added I18n suport.

At first glance one could argue this has duplicate language and could be
more concise, but I found it difficult to follow the GitLab best
practices for externalization (particularly "splitting sentences")
in it's previous implementation.

I also felt like although the code was more concise before, it was
fairly difficult to parse, and more difficult to extend.

* backfill license_helper_spec
* refactor license_helper

I've followed the best practices outlined here:
https://docs.gitlab.com/ee/development/i18n/externalization.html#best-practices

This is part of:
https://gitlab.com/gitlab-org/growth/product/-/issues/102