Populate canonical emails
Background migration to generate a canonical email based
on the user's primary email. "Canonical" means the Agent
part of the email address omitting `.` and anything after
any `+`. Only applies to Gmail since they are a service
that allows `.` and ignores anything after `+` in the Agent
and all variations arrive in the same inbox.