Geo Replication - Refactor action_spec
Currently this unit test was behaving
simlar to an integration test by
testing the axios calls.

Instead we decided it made more sense
to test the axios calls in the api file
and then just mock the function
calls in the action spec.

This MR converst the axios mock to ee/api
method mocks.