Add gauge metric global search queue size
This metric is used to track the size of updates that need
to be synchronized to Elasticsearch. Since this is a custom queue and
we're not using Sidekiq we needed to implement our own custom monitoring
for it.

This MR just sets a gauge every minute in a cron worker which will
ultimately be available in Prometheus. Later we will build a chart for
this.

See https://gitlab.com/gitlab-org/gitlab/issues/207095