Geo Replication - Rename Files
This MR is an effort at MVC.

This MR splits off from:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/26410

The overall focus of this and following related
changes is to replace Geo_Designs with
Geo_Replicable.

This MR renames all the files from
geo_designs => geo_replicable.

This MR also removes any "leftover"
references to the word "design".

This MR is the final MR to fully
split up the MR mentioned above.