Ensure errors are logged/tracked in repo import
Prior to this change, all errors that were raised in project import were
logged to `projection.log`, which is unstructured. This change starts
logging this information using the `track_exception` method so that we
log these errors to `exceptions_json.log`, which is structured and can
be more easily used by Kibana. It also sends the error to Sentry, so
that we can easily search for it there as well.

This will make it easier to debug errors that occur during our
repository imports.