Expose Enabled Replication to API
From the UI we currently do not know
if a replication is active or disabled.

Currently, we just show that nothing is synced.

However, as part of https://gitlab.com/gitlab-org/gitlab/-/issues/13309
we are going to differentiate between nothing synced
and synced disabled.

This adds the ability to do that by exposing the data through the API.