Track import job id by import state model
We are adding another import state tracking model for
Jira importer and we need to be able to track job id
of the importer. Refactoring this module so that it can
be reused with different project level import state
trackers by making use of the import state model instead
of the project.