Set props and listener explicitely
To have props and listeners' assignemtn consistent in the component,
this branch sets value and listener on <blob-content-edit>
explicitely instead of implicitely via v-bind="$attrs" and
v-on="$listeners".

Allows the component to not have content passed to it by default
This is important for the cases when the component is used in
creation of a new snippet.