Remove unused store module
As part of [Update dependencies table UI][1], the Dependency List is
being reduced from _two_ tabbed lists to just one.

That work is taking place behind the `dependency_list_ui` feature flag.
When it's disabled, the old UI will be rendered, which means the store
needs to be created with list modules for both the "All" and
"Vulnerable components" tabs.

When the flag is _enabled_, only the default "All" list module needs to
be set up. Otherwise, the UI would make an initial unnecessary network
request for the "Vulnerable components" tab/list, which isn't being
rendered.

[1]: https://gitlab.com/gitlab-org/gitlab/-/issues/195928