Change limited_ids_cached to in memory
This switches from storing the list of ids in a Redis key to storing
them in an in memory Set that is stored on a class instance variable.
This was done due to the fact that the list of ids was growing very
large and the cost of sending it back and for with Redis and
unmarshalling and then looking up the existence of an id in the list
was becoming considerable. Now in order to invalidate the cache across
nodes we use a last modified timestamp in Redis and thus our traffic
back and forth to Redis is much less.

See https://gitlab.com/gitlab-org/gitlab/-/issues/214440 for more
information.