Do not cast project IDs to string in API response
Previously we were casting the invalid project IDs to strings because
we were trying to cover the case where the client sends an invalid ID
to the backend. This was causing problems on client application because
it was expecting the IDs to be integers. I do believe that the effort
of covering the case where client sends an invalid ID will cause more
problems than it solves therefore I am removing the related logic to
assume the client is always sending us IDs as integers.

/closes #196729