Merge Geo Replication Views
Currently, everytime we add a new
replicable type, it requires a new item
in the side-nav.

This creates a new design pattern to allow us to better use space
to differentiate replicables.

This view contains all replicable types in one
consistent place.

This opens the door for further UI expansion as we open
up Geo Self-Service to countless new replicable types.