Move Design Management services to FOSS
This change is part of
https://gitlab.com/gitlab-org/gitlab/-/issues/212566 to move all Design
Management code to FOSS.

This MR moves the services and their specs.

It also ports some callbacks from `Noteable` which were added for the
benefit of Design Management, and were therefore in the ee directory.
These have been ported in this MR because of
design_user_notes_count_service_spec which tests the cache invalidation
and requires these callbacks.