Enable the `in this group` action in the Search dropdown
Previously, you would have to be on a Group's page to see this action
available.

With this commit, the action will also be present when a
project is owned by a Group.

This also fixes an issue where the dropdown would close whenever there
are no results from the autocomplete, which defeated the purpose of
having these actions in the first place.