Support resetting approval rules to defaults
A new `reset_approval_rules_to_defaults` param that will be sent
by the frontend when approval rules are set to be reset to project
defaults.

When that param is set to `true`, the backend will delete all
existing `ApprovalMergeRequestRule` of the merge request. It will
also append inapplicable rules so it'll behave like creating rules
on MR creation.