Preload commits markdown cache
This is for improving the performance of
Projects::MergeRequestsController#commits action.

Prevents n+1 requests to Redis when getting the value from cache
on render if the cache is warm.

Previously, it'll make an HMGET request to Redis per commit which
can result to 100 requests when viewing the commits in a MR.

Now, a single Redis pipeline is used via `with_markdown_cache`.