Update Prometheus chart to 10.4.1
We need a chart that will work with K8s versions 1.12 to 1.16. K8s v1.16
removed some deprecated APIs, which causes kube-state-metrics
(a sub-chart of Prometheus) to throw errors (in pod logs).

Prometheus chart v10.4.1 uses kube-state-metrics v1.9.1, which seems to
be compatible with K8s 1.14 to 1.17 (no errors in logs). We are unable
to test compatibility with K8s 1.12/1.13 since these are no longer
available on GKE.